\name{contamDE-package}
\alias{contamDE-package}
\docType{package}
\title{
 Differentially expression analysis for contaminated tumor samples and normal samples.
}
\description{
This package conducts statistical differential expression (DE) analysis between tumor samples and normal samples using RNA-seq count data from contaminated tumor samples and normal samples. The tumor samples could be either  matched or unmatched with normal samples.

}
\details{
\tabular{ll}{
Package: \tab contamDE\cr
Type: \tab contamDE\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-13\cr
License: \tab GPL (>= 2) \cr
}

}
\author{
Qi Shen, Jiyuan Shen, Hong Zhang

Maintainer: Hong Zhang <zhanghd@fudan.edu.cn>
}
\references{
SHEN Qi, HU Jiyuan, JIANG Ning, HU Xiaohua, LUO Zewei, and ZHANG Hong (2016) contamDE: Differential expression analysis of RNA-seq data for contaminated tumor samples. Bioinformatics 32(5): 705-712.
}

\keyword{ package }

