\name{prostate}
\alias{prostate}
\docType{data}
\title{
Prostate cancer data
}
\description{
Prostate cancer samples and adjacent normal cell samples were provided by 14 patients from Shanghai Changhai Hospital. This dataset contains the read count of 12,699 genes for each sample.
}
\usage{data("prostate")}
\format{
  A data frame with 12,699 observations on the following 29 variables.
  \describe{
    \item{\code{nameOfGene}}{Gene name}
    \item{\code{N1}}{Read count for normal sample of patient 1.}
    \item{\code{N2}}{Read count for normal sample of patient 2.}
    \item{\code{N3}}{Read count for normal sample of patient 3.}
    \item{\code{N4}}{Read count for normal sample of patient 4.}
    \item{\code{N5}}{Read count for normal sample of patient 5.}
    \item{\code{N6}}{Read count for normal sample of patient 6.}
    \item{\code{N7}}{Read count for normal sample of patient 7.}
    \item{\code{N8}}{Read count for normal sample of patient 8.}
    \item{\code{N9}}{Read count for normal sample of patient 9.}
    \item{\code{N10}}{Read count for normal sample of patient 10.}
    \item{\code{N11}}{Read count for normal sample of patient 11.}
    \item{\code{N12}}{Read count for normal sample of patient 12.}
    \item{\code{N13}}{Read count for normal sample of patient 13.}
    \item{\code{N14}}{Read count for normal sample of patient 14.}
    \item{\code{T1}}{Read count for heterogeneous tumor sample of patient 1.}
    \item{\code{T2}}{Read count for heterogeneous tumor sample of patient 2.}
    \item{\code{T3}}{Read count for heterogeneous tumor sample of patient 3.}
    \item{\code{T4}}{Read count for heterogeneous tumor sample of patient 4.}
    \item{\code{T5}}{Read count for heterogeneous tumor sample of patient 5.}
    \item{\code{T6}}{Read count for heterogeneous tumor sample of patient 6.}
    \item{\code{T7}}{Read count for heterogeneous tumor sample of patient 7.}
    \item{\code{T8}}{Read count for heterogeneous tumor sample of patient 8.}
    \item{\code{T9}}{Read count for heterogeneous tumor sample of patient 9.}
    \item{\code{T10}}{Read count for heterogeneous tumor sample of patient 10.}
    \item{\code{T11}}{Read count for heterogeneous tumor sample of patient 11.}
    \item{\code{T12}}{Read count for heterogeneous tumor sample of patient 12.}
    \item{\code{T13}}{Read count for heterogeneous tumor sample of patient 13.}
    \item{\code{T14}}{Read count for heterogeneous tumor sample of patient 14.}  
    }
}
\source{Ren S, Peng Z, Mao J, Yu Y, Yin C, Gao X, Cui Z, Zhang J, Yi K, Xu W, Chen C, Wang F, Guo X, Lu J, Yang J, Wei M, Tian Z, Guan Y, Tang L, Xu C, Wang L, Gao X, Tian W, Wang J, Yang H, Wang J and others. (2012). RNA-seq analysis of prostate cancer in the Chinese population identifies recurrent gene fusions, cancer-associated long noncoding RNAs and aberrant alternative splicings. \emph{Cell Research} 22(5), 806--821.}
\keyword{prostate cancer}
