\name{simcounts}
\alias{simcounts}
\docType{data}
\title{
A simulated count dataset.
}
\description{
Simulated RNA-seq read counts for eight patients with both normal and tumor samples. The eight patients belong to three tumor subtypes, i.e., four 'Type1' patients, two 'Type2' patients, and two 'Type3' patients.
}
\usage{data("simcounts")}
\format{
  A data frame with 10000 observations (read counts) on the following 16 samples.
  \describe{
    \item{\code{Type1Normal1}}{Read count for normal sample of the 1st 'Type1' patient.}
    \item{\code{Type1Normal2}}{Read count for normal sample of the 2nd 'Type1' patient.}
    \item{\code{Type1Normal3}}{Read count for normal sample of the 3rd 'Type1' patient.}
    \item{\code{Type1Normal4}}{Read count for normal sample of the 4th 'Type1' patient.}
    \item{\code{Type2Normal1}}{Read count for normal sample of the 1st 'Type2' patient.}
    \item{\code{Type2Normal2}}{Read count for normal sample of the 2nd 'Type2' patient.}
    \item{\code{Type3Normal1}}{Read count for normal sample of the 1st 'Type3' patient.}
    \item{\code{Type3Normal2}}{Read count for normal sample of the 2nd 'Type3' patient.}
    \item{\code{Type1Tumor1}}{Read count for tumor sample of the 1st 'Type1' patient.}
    \item{\code{Type1Tumor2}}{Read count for tumor sample of the 2nd 'Type1' patient.}
    \item{\code{Type1Tumor3}}{Read count for tumor sample of the 3rd 'Type1' patient.}
    \item{\code{Type1Tumor4}}{Read count for tumor sample of the 4th 'Type1' patient.}
    \item{\code{Type2Tumor1}}{Read count for tumor sample of the 1st 'Type2' patient.}
    \item{\code{Type2Tumor2}}{Read count for tumor sample of the 2nd 'Type2' patient.}
    \item{\code{Type3Tumor1}}{Read count for tumor sample of the 1st 'Type3' patient.}
    \item{\code{Type3Tumor2}}{Read count for tumor sample of the 2nd 'Type3' patient.}
    }
}
\keyword{simulated RNA-seq read counts}
